+++ src/server.c
@@ -624,7 +624,7 @@ static void ws_invoke_onmessage(void *arg) {
 
   SEXP data;
   if (info->textframes) {
-    PROTECT(data = Rf_ScalarString(Rf_mkCharLenCE((char *) buf, len, CE_UTF8)));
+    PROTECT(data = Rf_ScalarString(Rf_mkCharLenCE((char *) buf, (int) len, CE_UTF8)));
   } else {
     PROTECT(data = Rf_allocVector(RAWSXP, len));
     memcpy(NANO_DATAPTR(data), buf, len);
@@ -683,7 +683,7 @@ SEXP rnng_http_server_create(SEXP url, SEXP handlers, SEXP tls) {
   }
   srv->xptr = R_NilValue;
   srv->prot = R_NilValue;
-  srv->handler_count = handler_count;
+  srv->handler_count = (int) handler_count;
   for (R_xlen_t i = 0; i < handler_count; i++) {
     srv->handlers[i].callback = R_NilValue;
     srv->handlers[i].on_open = R_NilValue;
